RavenDB-17301 : add WaitForDatabaseToBeDeleted in RemoveDatabaseNode, since waiting on index+1 might not be enough
